Remove duplicate failure output in text UI
Individual propagation failures are displayed to the user in the UI
(both text and graphical). The failure output in the Transport module
looks more like it could be a debug output or intended for the log only.
In the GUI this is not a problem because output to stdout/stderr is
suppressed. In the text UI it is duplicating information and it can
generate a lot of output when there are many failures.

Make it write in the log only as otherwise the information might go
missing when using the GUI.