Add graceful stop in repeat mode - SIGUSR2
The repeat mode is basically an endless loop. It will run until the
process is terminated or a fatal error occurs. Currently the only way
for the user to stop repeat mode is to interrupt the process by pressing
Ctrl-C or sending a signal that terminates or kills the process. This
works but is not satisfactory if the delay between repeated syncs is
very short (using fsmonitor on a busy root, for example). The chance of
interrupting the process cleanly between syncs becomes very slim.

Add a new way to stop repeat mode and do so cleanly. A stop can be
requested (in this commit the request method is the signal SIGUSR2) by
the user at any time and Unison decides when it is safe to act on that
request. Any ongoing sync is allowed to run to completion. This way it
is guaranteed that the process terminates between syncs.