Add graceful stop in repeat mode - stdin
Add a new stop request method to terminate repeat mode. See commit
e72c1c5838ba6e85bcfba690a5a8fb228633a83a.

Monitor stdin to read any of the several stop conditions. Using stdin
for checking the repeat mode stop request opens up a wide range of
possibilities for the user, including manual and automated, local and
remote.

This method is usable only if stdin is an interactive terminal or is
redirected from a readable source that is not a regular file.