Prevent ssh process from receiving SIGINT
Certain signals can be generated by the terminal based on user input
(for example, Ctrl-C -> SIGINT). The generated signals are sent not just
to the process the user considers to be foreground but to all processes
in the foreground process group. As a consequence, if one process needs
the other(s) to do cleanup before terminating, it's no longer possible.
This happens to Unison when it tries to do cleanup with the remote
server. The ssh process has already died and the remote connection is
broken.

To prevent the ssh process from terminating too early, block some
signals (here, SIGINT) before spawning the ssh child process. Child
processes inherit the signal mask and signal settings so blocking some
signals before spawning effectively blocks these signals for the ssh
process (unless it changes the sigmask for itself).