Make [Uicommon.repeat] a structured preference
Parse and validate [Uicommon.repeat] at preference parsing instead of
letting the code using the value interpret it freely and possibly error
out very late in the process.

Restructure the code in text UI to better align with this new preference
representation and make it easier to expand in future.