CI.yml: Drop old ocaml versions on all but ubuntu
Drop ocaml 4.08 to 4.13 on macos and windows, on the theory that
building on ubuntu is sufficient to show portability to older ocaml,
and because windows builds in particular are too flaky.  With the new
scheme, it is possible that there might be an undetected issue with
older ocaml in OS-specific code, but basically I'd rather take that
risk than have to intervene in most CI runs.

Keep 1 windows build for i386, and one for x86_64.

Do not publish builds of old ocaml versions; there are no good reasons
to run them, and we are trying to be minimal in resource usage.

Limit to one dune build each on ubuntu/macos familes, since it's only
proving that the dune machinery is ok, not that the underlying code is
ok.

Reduce abicheck builds a bit.  There's no need to check 2.52.0 after
checking 2.52.1, and earlier versions can be checked on one OS version
or one ocaml version.  While this theoretically reduces test coverage,
our resource usage needs slimming, especially as the current
requirements become a baseline for a new CI system.