Differentiate Lwt_unix platform at run time
Lwt_unix does not implement all functions on Win32. Make it possible to
differentiate the Lwt_unix implementation at run time to avoid calling
unimplemented functions. Previously this differentiation was done by
probing the compiler OS platform (Win32? with Cygwin?) but this is not
reliable because it does not indicate the actual Lwt_unix implementation
that's linked in.