GHA: Improve CI run times on Windows
Add a very hacky way of speeding up CI builds on Windows. A huge chunk
of time (often just 5-10 minutes but could take 30-45 minutes) was spent
on building fontconfig cache, which... is absolutely not needed.
Building the cache is part of Cygwin installation and is not easy to
skip; that's what necessitated the hack.