Remove the custom inotify exception
It's just not needed and gets in the way. In the original library it was
most likely created to avoid dependency on the Unix library, which in
this case is not a problem.

Actual errno value is also not always equal between platforms. With Unix
library we always get the correct error code. In the original library
this was not a concern because only Linux was targeted.
inotify-compatible interfaces are now available on BSDs, Darwin (macOS)
and illumos-based OS.