Separate profile init and remote connection init
Establishing the remote connection does not need to be a strict part of
the profile init function. Separating those two tasks brings a few
benefits. The primary goal of this change is that handling of errors can
now be different depending on what exactly failed. Errors in profile
init are typically fatal and permanent. Errors in opening a connection
can be temporary. Profile init only needs to be done once. Establishing
the remote connection, on the other hand, can be attempted several times
once the profile init succeeds.

Reuse the [refreshConnection] function to be responsible for opening
connections whether it is the first time after profile init or
re-establishing a broken connection.