Make repeat mode slightly fault tolerant
With recent series of patches ensuring proper state and resource cleanup
is done at fatal exceptions, it is now possible to continue the repeat
mode even when receiving (some of) these exceptions. This allows the
repeat mode to survive the server temporarily going away (by it crashing
or the network going down, for example) or the fsmonitor helper process
crashing and other similar situations that can be reasonably expected to
be temporary.

There is a set of critical exceptions that will still cause the process
to exit. While it is theoretically possible (and often even safe) to
continue even at these critical errors, it is more likely that these are
a sign of programming error or an issue with the operating environment
(not enough memory, incorrect stack limit, ...).