Allow substrings in reconcile override preferences
Preferences to override reconcile results (force*, prefer*, nodeletion*,
noupdate*, nocreation*) currently require their value to be the complete
and exact root specification of one of the roots. This can be quite
tedious and redundant as the root strings can be very long.

As these preferences only need to _identify_ one of the roots, this
patch enables the preference value to be a substring of the root
specification. The substring must uniquely identify exactly one of the
roots.

To identify a root, the following _unique_ substrings are allowed:

 - an exact prefix or suffix of any length of a root's path

 - an exact prefix of any length of a remote root's hostname