Fix a potential future issue with repeat pref
Commit 5d6ab3cb9921a476e73e42307d40addc1b86f30a introduced a potential
issue with backwards compatibility. The issue could be triggered by a
future change in the syntax of the 'repeat' preference.

Due to legacy reasons, the value of the 'repeat' preference is sent to
the server, even though it is not used by the server. This means that
extending the set of valid values for the preference could break older
servers. This was not a problem before commit
5d6ab3cb9921a476e73e42307d40addc1b86f30a because the value of this
preference, even though sent to it, was never parsed by the server.
Since that commit, the value is however always parsed, and could cause
the server to raise a fatal exception if it fails to parse the new
syntax.

Fix the issue by capturing the error in the type instead of directly
raising it as an exception.