Avoid problems with quoting arguments to copyprog
Use Unix.open_process_args_{full,in} instead of Unix.open_process_{full,in},
passing command-line arguments as an array, so they are not parsed by the
shell and do not need to be quoted.

Add the function runExternalProgramArgs to src/external.ml to implement
this.

This is motivated by recent changes to rsync:
https://download.samba.org/pub/rsync/NEWS#3.2.4

This obviates the need for the copyquoterem preference, which is ignored for
backwards compatibility, but its documentation is removed.