Use accelerated copy functions with rsync
On filesystems that support block-level copy-on-write links (so-called
reflinks), the accelerated copy functions can be used by the rsync
algorithm to copy the unchanged data from the original file without
actually copying it.  If reflinks are not supported then an in-kernel
copy is attempted instead. Worst case, if this is not supported then
data will be copied as usual.

For large files with only few changes, this means the following (when
filesystem support exists):

 - syncing will be much faster; copying possibly gigabytes of
   data around can become almost free;

 - storage is only required for changed blocks; filesystem snapshots
   will record only changed blocks, as expected.

This has similar performance to rsync's "--in-place", but because a new
file with references is created, rather than modifying the existing
file, the atomicity guarantee provided by rename(2) is not compromised.