Increase max length of status messages in uitext
Most modern VT100 terminal emulators are able to switch automatic
line-wrapping off and on (the default state) with control sequences
ESC[?7l and ESC[?7h, aka reset and set mode DECAWM.

DECAWM and these control sequences are claimed to be supported by:
 - xterm
 - konsole
 - vte (GNOME terminal, XFCE terminal, ROX terminal and many more)
 - kitty
 - alacritty
 - rxvt and derivatives
 - Terminal.app
 - iTerm2
 - mintty
 - Windows Terminal (Windows 10+)
 - ANSI.SYS (sic!)
 - GNU screen
 - tmux
 - Linux console
 - FreeBSD console
 - NetBSD console (wscons)
 - and hopefully all other VT100/VT102 emulators, which is mostly
   anything used widely nowadays.

Some emulators may not wrap lines at all, providing horizontal scrolling
instead.

Terminal emulators known to not have this capability or using different
control sequences are the consoles on Solaris and illumos-based OS.

This patch leverages this terminal capability to provide status
information during updates scanning with length up to the maximum width
of the terminal (was previously capped at < 80) without wrapping to next
line(s). Code for clearing out the previous status message must be
updated too because even the clearing action can cause line(s) to be
wrapped (as it's just outputting a string of spaces).

This could be seen as an intermediate solution because it will basically
just chop off everything that does not fit on one line.

The previous method of truncating status information from the middle
continues to be used when the terminal is known to not support these
control sequences or when the preference 'dumbtty' is true.