Add a minimal terminal escape sequence parser
Replace the previously used regexp-based solution which was not only
silly but also only covered a tiny bit of all escape sequences.

The intention is not to produce an actual terminal emulator, rather to
clean out escape sequences for human presentation. A few of the escape
sequences will be interpreted, such as move cursor right (in this
interpretation it becomes a space character).