Improve remote shell terminal user interaction GUI
Make the uigtk3 password dialog properly asynchronous, and simplify and
improve the remote shell pty read-write code accordingly.

The user will now see output from the terminal continuously. Also user
input is now collected and written to terminal asynchronously. Now the
user can have a much better idea of what is going on with the remote
shell and when input from user is expected.

An open password dialog will not block establishing the connection and
starting updates scanning in the background. If the connection is
established (or broken) while the password dialog is still open then it
becomes just a message box displaying the output from terminal, without
requesting user input. Previously, any output (including everything that
does not require user input) from terminal was displayed to the user
together with an input field, which was not only confusing to the user
but blocked the entire GUI.

The pty read-write code is now much more robust and better sequenced,
not relying on timing so much (still a bit, but not in a too critical
way). Overall, the code is now more readable and should make sense more
easily.