Correctly wait for all sync transfers to be done
[Transport.run] can return before the entire transport queue is
processed. This is normal and expected. Marking dispensing as completed
immediately after [Transport.run] returned, on the other hand, was a
bug. This bug never became a real issue because the transport task pool
was constantly kept busy so that there were always outstanding tasks to
be waited on. Except for one corner case... when maxthreads is set to 1.
Then also the task pool size is 1 and there are no outstanding tasks
between one task completing and another one starting, which would
incorrectly consider all tasks to be done.