Remove STATIC from Makefiles (LDFLAGS works)
Static linking has become rather useless in general and is usable under
very narrow constraints only. There is no point in having a separate
STATIC variable controlling the build process. It's as easy (if not
easier if it's already in your environment) to set LDFLAGS=-static
instead of setting STATIC=true.