Fix an incorrectly split arg to exec
A change made to [Util.splitIntoWords] in 2018 turned every backslash in
the string into an escape character. This broke every string parameter
that would normally contain a backslash, like a path in Windows.

This patch fixes this only for argv[0] when executing ssh by not sending
argv[0] through the string splitter. This value could contain both
spaces and backslashes so it must not be split by spaces nor get the
backslashes mangled. The value comes from the 'sshcmd' user preference,
so it is intended to be a single non-splittable value. All arguments are
provided separately in the 'sshargs' preference and must not be included
in 'sshcmd'.