Stop redirecting stderr in Windows conpty
This is a workaround for an issue with Cygwin not detecting the
controlling terminal.

When using Windows pseudoconsole (conpty) to start Cygwin-linked ssh
(this is not a problem with native Windows port of ssh), Unison
redirects the child's stdin, stdout and stderr for its own use. Conpty
provides a controlling terminal, but Cygwin apparently checks for
existence of a controlling terminal only by checking stdin, stdout and
stderr. These all being redirected, Cygwin decides there is no
controlling terminal and does not create /dev/tty (likewise, all POSIX
API functions related to controlling terminal will fail).

OpenSSH reads the interactive auth password from /dev/tty. Since there
is no /dev/tty, there is no interactive auth and the result is a failed
connection.

The Windows port of OpenSSH works precisely because there is no /dev/tty
in Windows. The port has to directly use the Windows console anyway
(which in this case is provided by conpty, it's only Cygwin that doesn't
check for it properly).

This is an issue specific to using conpty. Before introduction of
conpty, child's stderr was redirected to [Unix.stderr]. In other words,
it was connected to the Windows console. Cygwin detects existence of a
controlling terminal on any of stdin, stdout, stderr so that was
sufficient. The new implementation did redirect stderr to the Windows
pseudoconsole but to Cygwin this looks like a pipe, not like the native
Windows console.

We don't actually want to redirect stderr in the first place. We want
stderr redirected to the pseudoconsole because then we can properly read
output from ssh together with all error output correctly interleaved.

It turns out that we can redirect stdin and stdout while leaving stderr
to be populated by Windows. Setting si.StartupInfo.hStdError to NULL has
Windows attach it to the pseudoconsole for the child process (that would
be ssh) _even when_ STARTF_USESTDHANDLES is set. This way we still get
stderr pointing to where we want it to point and as a bonus it looks
like a real console to Cygwin, even though it points to the exact same
target.

There is one immediately obvious side-effect. User input is echoed to
the console output and thus displayed back to the user in GUI. This is
not a major issue though and, importantly, it does not affect password
entry because OpenSSH is wise enough to turn off echoing for that.