Unify usage of Sys.unix, Sys.win32 and Sys.cygwin
Cygwin is not Windows and should not be considered as such by the code.
Cygwin is intended to be an almost-POSIX environment within Windows. By
default, code that must distinguish between POSIX/Unix and Windows must
treat Cygwin as POSIX.

There are a few cases where Cygwin can or may behave more like Windows,
not all of which are guaranteed to remain the same in future versions of
Cygwin. The most prominent example is path handling. Currently Cygwin
supports Windows-style paths in addition to POSIX paths, and this is
reflected in Unison code.

This patch combines two changes. First, it replaces all occurrences of
[Util.osType], which combined Windows and Cygwin into one, with standard
library's [Sys.os_type] and its shortcuts [Sys.win32], [Sys.cygwin] and
[Sys.unix].

Second, existing checks for Windows are in some, but not all, cases made
explicitly Windows-only, excluding Cygwin. This applies to all cases,
except Windows-style path handling and a few others where the code was
kept unchanged for backwards compatibility reasons. Ideally, in future
there should be no separate checks for Cygwin and Cygwin should follow
POSIX-style path handling only.