Rename inotify fsmonitor
Even though inotify is a Linux interface, there are other OS which
provide a compatible implementation (BSDs, even Darwin, and some
illumos-based OS) and may not have a native fsmonitor implementation of
their own.

Rename the inotify fsmonitor to make it clear that the code is universal
and not specific to Linux.