Fix a ctime fastcheck miss
There is a bug when using ctime for fastcheck, not unlike the bug with
mtime fixed by commit e2ac12466af64d5467c35703eae9b647bb092644.

The cause of the issue is similar to the one described in the
abovementioned commit. There is a problem when rescanning the extended
properties (like xattrs) of previously synced files (for example, when
changing the "xattrs" preference from false to true, which causes a
one-time rescan of all properties). If new extended properties are
detected but the file is eventually not synced (due to an error, or the
user actively skipping the file), these extended properties are no
longer detected on the next update check because the ctime in the
archive matches the ctime on disk and a complete rescan is this time not
requested (as it was a one-time rescan).

This patch fixes the bug by always resetting the ctime in the archive
when a one-time rescan is requested and new extended properties are
detected. Resetting the ctime in the archive guarantees that the
extended properties for the file are rescanned every time until they are
either successfully synced or the user preferences change (for example,
the "xattrs" preference is set to false again).