Force rescanning ext props when predicates change
Directory entries are re-read when certain predicates change due to the
list of synced files potentially changing. (Not reading directory
entries is one of the fastcheck mechanisms.) Change in these predicates
also affects ctime-based scanning of extended file properties (another
fastcheck mechanism).

When the list of synced files has potentially changed then extended file
properties (xattrs, for example) must also be rescanned because some
files that are newly included in the sync may already exist in the
archive. Otherwise, some updates for such files may be missed in case
ctime in the archive is the same as on disk but prior syncs had been
performed with preferences that disable syncing of extended file
properties.

This situation can happen, for example, by syncing some files without
xattrs (for example), then syncing some files with xattrs (for example)
while ignoring some other files, then finally again syncing the
previously ignored files but this time with xattrs (for example). These
files were already in the archive but without xattrs; any xattrs on
these files (unless added later so that ctime has changed) may now be
missed by the latest sync if ctime in archive is the same as that on
disk.