Store certain archive properties per path
Some of the predicates and preferences are stored in the archive as
archive properties. This is because the fastcheck feature needs to know
when these preferences change to potentially temporarily disable some of
the fastcheck mechanisms.

Previously, all these preferences were stored in the archive in one
copy, as if applying to the entire archive. This is incorrect because
subpaths within the same archive can be synced with different
preferences. This is achieved by using the "path" preference.

While clearly a bug, this was not really an issue so far because there
was only one fastcheck mechanism that relied on storing archive
properties like this, and it failed in a safe way. The only side-effect
of this "safe failure" was that directory entries may have been re-read
more often than with a fully working fastcheck. This would not be
noticed by users, unless syncing huge directories (hundreds of
thousands, millions of entries) or disks with really high latency
(networks mounts or really slow spinning disks).

With the new ctime-based fastcheck mechanism, this became a real issue
because this mechanism works file-by-file (fast directory checks force a
global rescan, for comparison). This means that syncing different
subpaths in the same archive with different preferences will eventually
cause missing some extended file properties (xattrs, for example) as the
fastcheck mechanism gets confused by preferences that really apply to a
different subpath.

There are very specific conditions under which this can happen, though,
so most users should not be impacted:

 1. subpath A is synced without some extended properties;
 2. subpath B is synced with some extended properties (this signals a
    preference change to the fastcheck mechanism, temporarily disabling
    ctime-based mechanism);
 3. subpath A is synced with same extended properties as step 2 (as far
    as fastcheck mechanism is concerned, there is no preference change
    during this sync);
    now, extended properties on files in subpath A _that have not
    changed since the first sync_ (ctime has not changed) can be missed
    (as there was no preference change, the ctime-based fastcheck
    mechanism is active instead of being temporarily disabled).

This patch stores certain archive properties per subpath they apply to,
instead of globally. The solution is currently very trivial but at the
moment there is no cause to refine it more as most users wouldn't have
thousands of paths and/or predicates with a single archive.

Note that this patch intentionally does not fix the non-optimal
re-reading of directory entries that can happen under these conditions.