Help [Unix.symlink] create Windows dir symlinks
Windows differentiates between file symlinks and directory symlinks.
Correct type must be given when creating a symlink. Unison currently
does not keep this information on symlinks, so all new symlinks are
created as file symlinks by default. However, there is a heuristic in
[Unix.symlink] which checks the link target to see if it's a directory
or not. Obviously, this only works for existing targets. Moreover, it
only works for absolute path targets, unless the symlink is created in
the current working directory.

This patch helps the heuristic correctly discover the type of relative
path targets by first changing working directory to where the symlink is
being created.