Use UTF-8 encoding-safe string truncation in UI
The names of roots can be very long and are truncated in UI in the list
of updates. Truncating by number of bytes does not play well with UTF-8
input. This patch makes the truncation safe on UTF-8 encoded strings
while keeping the length of truncated strings roughly as before (and the
result should also look correct for Western languages, maybe also for
other scripts).