Improve display of cmdline argument parsing errors
Do not include the list of all command line options (which by now is
very long) with an error message when validation of a parsed argument
value fails.

This patch only affects errors from more complex validation of certain
preferences, such as 'ignore' predicates. Such errors are produced very
rarely to begin with - more common errors, such as unknown argument,
invalid argument type (string, instead of a number), are still displayed
as before. It may make sense to dispaly a list of all options with an
"unkown argument" error but it does not make sense with more complex
logic errors.

Without this change, some of the command line argument validation errors
end up in the GUI, displaying the long list of options in an overflowing
message box, with the actual error obscured or not visible.