Remove unneeded [Fingerprint.string]
Usage of [Fingerprint.string] in archive names is incorrect - archive
names are based on MD5 which may or may not be the same as the file
fingerprinting algorithm.

The other use of [Fingerprint.string] is in debug output of tests, where
it's used to debug the test framework itself. Remove this usage for now.

As such, [Fingerprint.string] becomes unused and can be removed.