Ignore force = newer/older if mtimes are equal
force = newer/older preference works based on the difference of mtime in
each replica. If mtimes are equal then it always propagates from root 2
to root 1. This yields unexpected results when propagating changes that
don't update the mtime (that's basically all metadata updates), with
different roots treated clearly differently.

This patch makes it explicit that newer/older are only defined for when
mtimes differ, ignoring the force = newer/older preference (and other
related preferences, such as prefer and *partial) if mtimes are equal.