Stop using replica paths in connection management
Registering and looking up remote connections was mistakenly considering
replica paths present in clroots (and also in Common.roots, but that's
irrelevant at the moment). This is not correct because a connection is
not specific to a path and the path is not used in opening or managing
the connection. This in turn can cause multiple connections to be opened
when a single connection could be reused instead.