Execute less code when retrying after an error
In text UI, when retrying after an error in repeat mode, unnecessary
amount of code was being re-executed on each retry. This patch removes
the early init code from the retry loop as it only needs to be executed
once and will either succeed or fail, there is no retry.

(best reviewed with whitespace ignored)