Display very early errors in GUI
Previously, errors during very early GUI init process, before the main
window was created, would not be displayed in the GUI. The user would
not know what was happening because the GUI would just not start. The
only way the user would see the error was to read stderr, if available.

This patch makes it possible to display very early errors in GUI, as
long as GTK init has been successful. If the GTK init itself fails then
the user still must have access to stderr to understand what happened.