doc: Adjust INSTALL.md overall instructions
Change "package repositories" to "packaging systems".  Add Debian as
an example.  Caution that packaging systme bugs should be filed with
packaging systems.

Point to docs of prereqs, and do not attempt to open-code ocaml
instructions.  Fixes #966.

Change Unix-like to "mostly POSIX".

Note that a C compiler is needed; be vague for now until someone tries
with one that is too ancient, because no maintaened system should be
problematic.