Fix incorrect retry when SIGINT reaches ssh
Make sure that the retry loop is never activated if a SIGINT has been
received. In some cases the SIGINT could have been incorrectly detected
as a lost connection which would normally enable the retry loop. This
can happen when the ssh process receives SIGINT as part of the process
group and shuts down, which to us looks like a lost network connection.