Change props of the temporary file for diffing
For purposes of running a local diff, a remote file is copied by the
same routines that are used fo the actual sync. This includes all checks
and verifications. Using "safe" dummy props (as is currently done) to
copy the remote file into a temporary file can in some cases trip those
checks. This mainly happens when permissions are influenced by
fs/storage setup or ACL inheritance, for example.

This patch sets the temporary file props to be the same as what the
local file already has. The hope is that setting these props would
succeed during a normal sync, so it should also work for the temporary
file.