Add a specific message for Stack_overflow exception
Stack_overflow is quite certainly a result of programming error. Even
so, you would rarely see it and receive a SIGSEGV first (except perhaps
with bytecode runtime). OCaml since version 5 is allocating a separate
stack and thus it is quite likely to see a Stack_overflow exception
without crashing the entire program.