INSTALL: simplify and make consistent GNU make
The basic issue is that on BSDs at least, "make" is not GNU make.
On Debian GNU/Linux, GNU make is available as gmake, and this seems
normal.  Thus, use "gmake" as the example invocation of gmake,
consistently.  Leave the mac text as "make", as this is the only known
platform where "gmake" won't work.