INSTALL: rototill macOS section
Explain that macOS is a subset of mostly-POSIX, but that it's odd in
that normal tools and /usr/include are missing, and that the standard
approach is to install CLT or Xcode.

More or less, I think that if someone has a mac on which they can
build other POSIXY programs, unison will build fine.  If other mac GUI
programs build, unison mac GUI should build also.  And if the mac is
not set up to build programs, it's not really about unison, so a
CLT/Xcode hint suffices.