props: Stop allowing 1-hour differences in times
This was introduced in adf68c98d042bcc484fbd2e5dfbdc97ca2ef1881 in
2009.  The theory is that on Windows with FAT, filesystem times are in
an undetermined timezone (a design defect) and as DST happens, the
times of files already there change.  While this is true, ignoring an
hour change also ignores valid changes on filesystems that store times
correctly (in UTC, as the major example).

Ignoring an hour is not enough; if the host's timezone changes by 3
hours (or anything not 1), then times in FAT move by that amount.
It's also too much, in that the workaround ignores time changes on
filesystems with a stable time representation.

Therefore, remove the workaround, and anyone who is using FAT,
changing timezones, and syncing times (expected to be very few) can
propose how to deal with this in a way that doesn't adversely affect
others.

Fixes #72.