GHA: Migrate to setup-ocaml v3
This is mainly required for building in Windows and for enabling OCaml 5
builds in Windows. Building in Windows has always been a bit tricky. For
many years, a separate opam repository with patched compilers (and
packages), together with (optional) pre-compiled compiler binaries was
used to make these builds possible. That separate repository, however,
was created and maintained by a single person who decided to retire it
as of August 2021.

After a while, a migration path was proposed, which started with a
compatibility mechanism in the upstream opam repository and concluded
with a recent release of opam 2.2.0, a version which supports Windows
natively. setup-ocaml v3 makes use of opam 2.2+ and is thus required to
complete the migration of Windows build environments.

Use of opam 2.2 and the upstream opam repository requires a few changes
in the CI build process:

 - The depext mechanism is no longer available in Windows. Not sure if
   this is permanent or not, but since Unison builds were not relying on
   it heavily anyway (just for GTK), work around this by explicitly
   installing GTK before the GUI builds.

 - Many (likely most) opam packages are not compatible with Windows,
   often in just missing little things in the build process. Luckily,
   Unison does not rely on packages other than lablgtk and its
   dependencies, which just require a workaround for pkg-config (see
   further comments in the build script).

 - ocaml-env is no longer required nor available. Any invocations can
   simply be removed from the build script.