GHA: Work around pkgconf issues
A recent dune release uncovered issues with pkgconf (required by cairo2
and lablgtk3) which break some builds. In the current workflow only
builds on ubuntu-20.04 are affected (MSVC GUI builds were also broken
but these are not yet merged into the workflow). Since we don't rely on
pkgconf, we can just remove it and let dune fall back to pkg-config.