GHA: Work around pkgconf issues (part 2)
A recent dune release uncovered issues with pkgconf (required by cairo2
and lablgtk3) which break some builds. In the current workflow, builds
on ubuntu-20.04 and with MSVC are affected. Since we don't rely on
pkgconf, we can just remove it and let dune fall back to pkg-config.
For MSVC builds we can go even further and try to remove Cygwin from
the build environment as much as possible.