Makefile: Remove local variables from linking rules
Some make implementations (OpenBSD make, NMAKE) do not support
target-specific local variables. This patch removes any local
variables that are used directly by the linking rules (as these
are now duplicated, local variables are no longer required).

The CAMLFLAGS variable is currently not removed (it is required for
compilation rules of the dependents) and will be handled separately
later on.