dev/ktrace-netbsd: Add script to find used bin-like progams
This script, for NetBSD only, aims to find bin-like programs invoked
during unison's build.

Because the mechanisms for asking the system about system calls are
very different on different systems, it's likely that a mostly-new
script would be needed for other systems, but we could also extend
with a case and rename, if that makes sense.  But, the hard part is
the last, non-portable, command.