Fix the default key display in TUI's [warnPrinter]
Define Enter explicitly as input in the TUI warning printer. This is not
a functional change, it only changes the default key that is displayed
to the user, to match the prompt.

Enter/Return as input is treated as the default selection, as if user
had pressed any of the default keys for the respective action (if there
are multiple default keys defined then they are all equivalent). The
first default input key is displayed to the user so they know what
action is the default.

In the warning printer the first default input key used to be <space>.
This was a bit confusing as the prompt reads "Press return to continue."
This patch defines Enter as the default input key in its own right,
while also making it the first default which gets displayed to the user.