Add option to sync files in new profile assistant
GTK file chooser allows selecting either files or directories but not
both. File choosers in the new profile assistant were configured for
selecting directories, with the expectation that this covers the vast
majority of usages. This patch adds an option for the user to choose
files instead of directories, enabling specifying single files as sync
roots.