Open fpcache file for writing only when needed
There is no need to truncate and write the fpcache file at every scan.
This patch changes the code to open the file for writing only when
needed. As a side effect, the previous cache file is kept intact until
there is new cache to be written (which may not happen at every scan).