Remove locks at connection close and process exit
Currently, when the process exits unexpectedly or a connection is
dropped between client and server, any locks files that are held may
remain held.

Try to clean up as much as possible by removing any held lock files
when connection is closed or when the process exits normally (that is,
not crashed or killed).