Add a mechanism for compiler compatibility modules
Compatibility with a range of compiler (and its standard library)
versions can in vast majority of cases be handled by writing code for
the oldest supported compiler version. Often times even breaking changes
can be handled by changing the code slightly, although in many cases
that means that older code will not be compatible with future compiler
releases.

This patch adds a simple mechanism for compatibility shims in the build
system for the rare cases when it is not possible or sufficient to write
single code covering all supported compiler (and standard library)
versions, including future versions.