Use [Digest.MD5] explicitly
The [Digest] module currently provides an implementation of the MD5
algorithm but this is implicit. Sub-modules for explicit algorithm
choice were introduced only starting with OCaml 5.2.0. There is a real
risk of the [Digest] module default algorithm in future being changed
to something other than MD5 (or the implicit default being removed
completely).

This patch makes the use of MD5 explicit to make the current code
compatible with future compiler releases should that change happen. This
patch has no effect when compiling with current compilers (or future
compilers if the default algorithm is not changed).