Reset preferences to defaults in server process
With the introduction of multi-version compatibility and the 'features'
mechanism, it is possible for a long-running server process to retain
"ghost" values to some preferences when clients with different features
connect. This happens when a preference value is set by a previous
client but not by the next client (due to it not supporting that
preference, for example). To fix this, simply reset the preferences
values on the server at every client connection.