Reuse the documentation window if already open
Currently, if the user does not close the documentation window, every
time documentation is requested (by a Help button or from the Help
menu), a new documentation window is opened.

With this patch, when the user requests help, if the documentation
window is already open, raise it to top and navigate to the requested
section.